# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ExplainRequest": {
        "additionalProperties": false,
        "properties": {
          "profileId": {
            "title": "Profileid",
            "type": "string"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          }
        },
        "required": [
          "profileId",
          "statement"
        ],
        "title": "ExplainRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObservationIn": {
        "additionalProperties": false,
        "properties": {
          "childNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Childnumber"
          },
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "estimated": {
            "additionalProperties": true,
            "title": "Estimated",
            "type": "object"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "measured": {
            "additionalProperties": true,
            "title": "Measured",
            "type": "object"
          },
          "profileId": {
            "title": "Profileid",
            "type": "string"
          },
          "source": {
            "default": "cursor",
            "title": "Source",
            "type": "string"
          },
          "sqlId": {
            "default": "",
            "title": "Sqlid",
            "type": "string"
          }
        },
        "required": [
          "profileId",
          "label"
        ],
        "title": "ObservationIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "OracleDBHarness is the shared execution, policy and copilot layer for Oracle work.\n\nEvery operation carries an operation ID, an actor, a target, typed parameters,\ncapability requirements, a risk class and limits, and produces an execution record\nwith a structured result and verification evidence.\n\nNotes that apply to the whole API:\n\n* Free-form SQL runs only on targets where an administrator enabled worksheets. A\n  keyword filter is not a read-only boundary, so the Oracle account is the boundary.\n* Production targets are observation only in this release.\n* Applying a copilot proposal changes an editor buffer. It never runs, compiles or\n  commits anything.\n",
    "title": "OracleDBHarness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tuning/explain": {
      "post": {
        "description": "Produce an estimated plan without executing the statement.",
        "operationId": "explain_api_v1_tuning_explain_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Explain Api V1 Tuning Explain Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain",
        "tags": [
          "tuning"
        ]
      }
    },
    "/api/v1/tuning/observations": {
      "get": {
        "operationId": "list_observations_api_v1_tuning_observations_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Observations Api V1 Tuning Observations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Observations",
        "tags": [
          "tuning"
        ]
      },
      "post": {
        "description": "Save a before or after observation so a change can be compared later.",
        "operationId": "save_observation_api_v1_tuning_observations_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Observation Api V1 Tuning Observations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Observation",
        "tags": [
          "tuning"
        ]
      }
    },
    "/api/v1/tuning/observations/compare": {
      "get": {
        "description": "Compare two saved observations, keeping estimates and measurements apart.",
        "operationId": "compare_observations_api_v1_tuning_observations_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "beforeId",
            "required": true,
            "schema": {
              "title": "Beforeid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "afterId",
            "required": true,
            "schema": {
              "title": "Afterid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compare Observations Api V1 Tuning Observations Compare Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Observations",
        "tags": [
          "tuning"
        ]
      }
    },
    "/api/v1/tuning/{profile_id}/cursors": {
      "get": {
        "description": "Find cached cursors. Reading them never re-runs the user statement.",
        "operationId": "search_cursors_api_v1_tuning__profile_id__cursors_get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqlId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sqlid"
            }
          },
          {
            "in": "query",
            "name": "textFilter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Textfilter"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Search Cursors Api V1 Tuning  Profile Id  Cursors Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Cursors",
        "tags": [
          "tuning"
        ]
      }
    },
    "/api/v1/tuning/{profile_id}/cursors/{sql_id}": {
      "get": {
        "description": "Statistics and, where the views are readable, the plan actually used.",
        "operationId": "cursor_detail_api_v1_tuning__profile_id__cursors__sql_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sql_id",
            "required": true,
            "schema": {
              "title": "Sql Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "childNumber",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Childnumber",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cursor Detail Api V1 Tuning  Profile Id  Cursors  Sql Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cursor Detail",
        "tags": [
          "tuning"
        ]
      }
    },
...
  }
}